@@ -1,53 +1,72 @@
 {
   "name": "express-mvc",
   "version": "0.0.0",
-  "private": true,
   "scripts": {
-    "start": "webpack-dev-server --mode development --hot",
-    "server": "node ./bin/www",
-    "client": "npm start --prefix client",
-    "dev": "concurrently \"npm run start\" \"npm run client\"",
-    "build": "webpack --mode production",
-    "transpile": "babel views -d dist --copy-files"
+    "start": "nodemon index.js",
+    "lint": "eslint app.js",
+    "pack": "webpack --mode development",
+    "babel": "babel src -d views",
+    "build": "npm run pack && npm start"
   },
   "main": "index.js",
   "dependencies": {
+    "axios": "^0.18.0",
+    "babel-polyfill": "^6.26.0",
+    "babel-register": "^6.26.0",
+    "body-parser": "^1.18.3",
     "bootstrap": "^4.1.3",
     "concurrently": "^4.0.1",
-    "consolidate": "^0.15.1",
     "cookie-parser": "~1.4.3",
     "debug": "~2.6.9",
     "express": "^4.16.4",
+    "fs": "0.0.1-security",
     "http-errors": "~1.6.2",
     "joi": "^14.0.1",
+    "jsonwebtoken": "^8.3.0",
     "module": "^1.2.5",
     "morgan": "~1.9.0",
     "multer": "^1.4.1",
     "mysql2": "^1.6.1",
     "npm": "^6.4.1",
+    "react": "^16.6.3",
+    "react-dom": "^16.6.3",
     "react-hot-loader": "^4.3.12",
     "react-redux": "^5.1.0",
+    "react-router-dom": "^4.3.1",
     "reactstrap": "^6.5.0",
     "redux": "^4.0.1",
-    "sequelize": "^4.41.0"
+    "redux-saga": "^0.16.2",
+    "redux-thunk": "^2.3.0",
+    "sequelize": "^4.41.0",
+    "webpack-node-externals": "^1.7.2"
   },
   "devDependencies": {
+    "autoprefixer": "^9.3.1",
+    "babel": "^6.23.0",
     "babel-cli": "^6.26.0",
     "babel-core": "^6.26.3",
+    "babel-eslint": "^10.0.1",
     "babel-loader": "^7.1.4",
+    "babel-plugin-transform-runtime": "^6.23.0",
     "babel-preset-env": "^1.7.0",
+    "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
+    "babel-preset-react-app": "^6.1.0",
     "babel-preset-stage-0": "^6.24.1",
     "css-loader": "^1.0.1",
+    "eslint": "^5.9.0",
+    "eslint-config-airbnb": "^17.1.0",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-jsx-a11y": "^6.1.2",
+    "eslint-plugin-react": "^7.11.1",
+    "extract-text-webpack-plugin": "^3.0.2",
     "file-loader": "^2.0.0",
     "html-webpack-plugin": "^3.2.0",
     "node-sass": "^4.9.4",
-    "react": "^16.6.0",
-    "react-dom": "^16.6.0",
-    "react-router-dom": "^4.3.1",
+    "prettier": "^1.15.1",
     "sass-loader": "^7.1.0",
     "style-loader": "^0.23.1",
-    "webpack": "^4.23.1",
+    "webpack": "^4.25.1",
     "webpack-cli": "^3.1.2",
     "webpack-dev-server": "^3.1.10"
   }
