@@ -1,52 +1,77 @@
-import React from 'react';
 import { hot } from 'react-hot-loader';
+import { renderToString } from 'react-dom/server';
 
-var createError = require('http-errors');
-var express = require('express');
-var path = require('path');
-var cookieParser = require('cookie-parser');
-var logger = require('morgan');
-var engines = require('consolidate');
+import React from 'react';
+import { Provider } from 'react-redux';
+import { StaticRouter } from 'react-router-dom';
+import App from './src/Components/App';
+import configureStore from './src/configureStore';
+import renderFullPage from './template/template';
 
-var userRouter = require('./routes/user');
+const createError = require('http-errors');
+const express = require('express');
+const path = require('path');
+const cookieParser = require('cookie-parser');
+const logger = require('morgan');
+
+const userRouter = require('./routes/user');
 const profileRouter = require('./routes/profile');
-var indexRouter = require('./routes/index');
+import userInfoRouter from './routes/userInfo';
 
-var app = express();
+const app = express();
 
 // view engine setup
-app.set('views', path.join(__dirname, 'views'));
-app.engine('html', engines.mustache);
-app.set('view engine', 'html');
-///////////////////////////////////
-///////////////////////////////////
+// app.set('views', path.join(__dirname, '/public'));
+//  app.set('view engine', 'js');
+//  app.engine('js', require('express-react-views').createEngine());
 app.use(logger('dev'));
 app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
-app.use(express.static(path.join(__dirname, '/public')));
+app.use(express.static(path.join(__dirname, 'public')));
 
-// app.use(express.static('./dist'));
-app.use('*', indexRouter);
-app.use('/',indexRouter);
-app.use('/api/user', userRouter);
-app.use('/profile',profileRouter);
-app.use('/api/users', usersRouter);
+app.use('/api/login', userRouter);
+app.use('/api/userinfo', userInfoRouter);
+app.use('/profile', profileRouter);
+
+app.use((req, res) => {
+  const context = {};
+  let initialState = {
+    userlogin: false,
+    toggle: false,
+    token: '',
+    userAdmin: '',
+  }
+
+  const store = configureStore(initialState);
+  const html = renderToString(
+    <Provider store={store}>
+      <StaticRouter context={context} location={req.url}>
+        <App />
+      </StaticRouter>
+    </Provider>,
+  );
+  const preloadedState = store.getState();
+  res.send(renderFullPage(html, preloadedState));
+  res.end();
+});
+
+
 
 // catch 404 and forward to error handler
-app.use(function(req, res, next) {
+app.use((req, res, next) => {
   next(createError(404));
 });
 
 // error handler
-app.use(function(err, req, res, next) {
+app.use((err, req, res) => {
   // set locals, only providing error in development
   res.locals.message = err.message;
   res.locals.error = req.app.get('env') === 'development' ? err : {};
 
   // render the error page
   res.status(err.status || 500);
-  res.render('error');
+  res.json({});
 });
 
 module.exports = hot(module)(app);
