@@ -1,28 +1,17 @@
-const express = require('express');
-const app = express();
-
-exports.login = (req, res) => {
-    res.json([{
-        id: 1,
-        name: "Hiccup",
-        password: 'hiccup'
-      }, {
-        id: 2,
-        name: "King Arthur",
-        password: 'king-arthur'
-      }]);
-    res.end();
-}
-
-exports.index = (req, res) => {
-  res.json([{
-    id: 1,
-    name: "Hiccup",
-  }, {
-    id: 2,
-    name: "King Arthur",
-  }]);
-res.end();
-console.log('The game started');
-}
-
+exports.rootIndex = (req, res) => {
+  const raw = `
+      <!DOCTYPE html>
+        <html lang="en">
+        <head>
+            <meta charset="UTF-8">
+            <title>Title</title>
+        </head>
+        <body>
+      
+        <div id="app"></div>
+        <script src="dist/bundle.js"></script>
+        </body>
+        </html>  
+    `;
+  res.send(raw);
+};
