@@ -1,13 +1,9 @@
 const express = require('express');
+
 const router = express.Router();
-const indexController=require('../controllers/indexController');
-const path = require('path');
+const indexController = require('../controllers/indexController');
 
-router.get('/customer', indexController.login);
+router.get('/', indexController.rootIndex);
+// router.get('/', indexController.index);
 
-router.get('/', indexController.index);
-
-router.get('*', (req, res) => {
-    res.sendFile(path.resolve(__dirname, 'public', 'index.html'));
-})
 module.exports = router;
