@@ -1,26 +1,27 @@
 const express = require('express');
+
 const router = express.Router();
 const multer = require('multer');
 
-var storage = multer.diskStorage({
-    destination: function (req, file, cb) {
-      cb(null, 'upload')
-    },
-    filename: function (req, file, cb) {
-      cb(null, file.fieldname + '-' + Date.now() + '.jpg')
-    }
-  })
-  
-  var upload = multer({ storage: storage }).single('profileImage');
+const storage = multer.diskStorage({
+  destination(req, file, cb) {
+    cb(null, 'upload');
+  },
+  filename(req, file, cb) {
+    cb(null, `${file.fieldname}-${Date.now()}.jpg`);
+  },
+});
 
-  router.post('/', function (req, res) {
-    upload(req, res, function (err) {
-      if (err instanceof multer.MulterError) {
-        // A Multer error occurred when uploading.
-      } else if (err) {
-        // An unknown error occurred when uploading.
-      }
-      // Everything went fine.
-    })
-  })
+const upload = multer({ storage }).single('profileImage');
+
+router.post('/', (req, res) => {
+  upload(req, res, (err) => {
+    if (err instanceof multer.MulterError) {
+      // A Multer error occurred when uploading.
+    } else if (err) {
+      // An unknown error occurred when uploading.
+    }
+    // Everything went fine.
+  });
+});
 module.exports = router;
