@@ -1,6 +1,9 @@
 const express = require('express');
+
 const router = express.Router();
-const userControllers=require('../controllers/userController');
+const userControllers = require('../controllers/userController');
+
+// router.get('/', userControllers.getAll);
 
 router.get('/:id', userControllers.userId);
 
